requirements.txt
Django>=4.0,<5.0
djangorestframework
drf-yasg
psycopg2-binary
django-import-export
openpyxl
pandas